+++ elex/cli/hooks.py
@@ -12,7 +12,8 @@ def add_election_hook(app):
         liveresults=not app.pargs.not_live,
         resultslevel=app.pargs.results_level,
         setzerocounts=app.pargs.set_zero_counts,
-        is_test=False
+        is_test=False,
+        raceids=[]
     )
 
     if app.pargs.data_file:
@@ -24,6 +25,9 @@ def add_election_hook(app):
     if app.pargs.local_only:
         app.election.national = False
 
+    if app.pargs.raceids:
+        app.election.raceids = [x.strip() for x in app.pargs.raceids.split(',')]
+
 
 def cachecontrol_logging_hook(app):
     """
@@ -31,7 +35,9 @@ def cachecontrol_logging_hook(app):
     """
     from cachecontrol.controller import logger
     formatter = logging.Formatter(LOG_FORMAT)
+
     for handler in app.log.backend.handlers:
         handler.setFormatter(formatter)
         logger.addHandler(handler)
+
     logger.setLevel(logging.DEBUG)