+++ tests/__init__.py
@@ -3,8 +3,7 @@ import unittest
 from elex.api import Election, DelegateReport, utils
 from time import sleep
 
-API_MESSAGE = "We require that you export AP_API_KEY in your environment in \
-order to test AP connectivity."
+API_MESSAGE = "We require that you export AP_API_KEY in your environment in order to test AP connectivity."
 
 
 class DelegateReportTestCase(unittest.TestCase):
@@ -17,6 +16,29 @@ class DelegateReportTestCase(unittest.TestCase):
             delsum_datafile=self.delsum_datafile
         )
         self.delegate_reports = d.candidate_objects
+
+
+class ElectionResultsParseIdsTestCase(unittest.TestCase):
+    data_url = 'tests/data/20151103_national.json'
+
+    def setUp(self, **kwargs):
+        e = Election(
+            datafile=self.data_url,
+            testresults=False,
+            liveresults=True,
+            is_test=False,
+            raceids=['7583']
+        )
+        self.election = e
+        self.resultslevel = e.resultslevel
+        self.raw_races = e.get_raw_races()
+        self.race_objs = e.get_race_objects(self.raw_races)
+        self.ballot_measures = e.ballot_measures
+        self.candidate_reporting_units = e.candidate_reporting_units
+        self.candidates = e.candidates
+        self.races = e.races
+        self.reporting_units = e.reporting_units
+        self.results = e.results
 
 
 class ElectionResultsTestCase(unittest.TestCase):